+++ internal/app/machined/main.go
@@ -23,6 +23,7 @@ import (
 	"github.com/siderolabs/go-cmd/pkg/cmd/proc/reaper"
 	debug "github.com/siderolabs/go-debug"
 	"github.com/siderolabs/go-procfs/procfs"
+	"go.uber.org/zap"
 	"golang.org/x/sys/unix"
 
 	"github.com/siderolabs/talos/internal/app/apid"
@@ -30,6 +31,7 @@ import (
 	"github.com/siderolabs/talos/internal/app/machined/pkg/runtime"
 	"github.com/siderolabs/talos/internal/app/machined/pkg/runtime/emergency"
 	v1alpha1runtime "github.com/siderolabs/talos/internal/app/machined/pkg/runtime/v1alpha1"
+	startuptasks "github.com/siderolabs/talos/internal/app/machined/pkg/startup"
 	"github.com/siderolabs/talos/internal/app/machined/pkg/system"
 	"github.com/siderolabs/talos/internal/app/machined/pkg/system/services"
 	"github.com/siderolabs/talos/internal/app/maintenance"
@@ -161,17 +163,9 @@ func runDebugServer(ctx context.Context) {
 	}
 }
 
-//nolint:gocyclo
 func run() error {
-	errCh := make(chan error)
-
 	// Limit GOMAXPROCS.
 	startup.LimitMaxProcs(constants.MachinedMaxProcs)
-
-	// Set the PATH env var.
-	if err := os.Setenv("PATH", constants.PATH); err != nil {
-		return errors.New("error setting PATH")
-	}
 
 	// Initialize the controller without a config.
 	c, err := v1alpha1runtime.NewController()
@@ -181,10 +175,35 @@ func run() error {
 
 	revertSetState(c.Runtime().State().V1Alpha2().Resources())
 
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	logger, err := c.V1Alpha2().MakeLogger("early-startup")
+	if err != nil {
+		return err
+	}
+
+	start := time.Now()
+
+	// Run startup tasks, and then run the entrypoint.
+	return startuptasks.RunTasks(ctx, logger, c.Runtime(), append(
+		startuptasks.DefaultTasks(),
+		func(ctx context.Context, log *zap.Logger, _ runtime.Runtime, _ startuptasks.NextTaskFunc) error {
+			logger.Info("early startup done", zap.Duration("duration", time.Since(start)))
+
+			return runEntrypoint(ctx, c)
+		},
+	)...)
+}
+
+//nolint:gocyclo
+func runEntrypoint(ctx context.Context, c *v1alpha1runtime.Controller) error {
+	errCh := make(chan error)
+
 	var controllerWaitGroup sync.WaitGroup
 	defer controllerWaitGroup.Wait() // wait for controller-runtime to finish before rebooting
 
-	ctx, cancel := context.WithCancel(context.Background())
+	ctx, cancel := context.WithCancel(ctx)
 	defer cancel()
 
 	drainer := runtime.NewDrainer()
@@ -237,7 +256,7 @@ func run() error {
 	initializeCanceled := false
 
 	// Initialize the machine.
-	if err = c.Run(ctx, runtime.SequenceInitialize, nil); err != nil {
+	if err := c.Run(ctx, runtime.SequenceInitialize, nil); err != nil {
 		if errors.Is(err, context.Canceled) {
 			initializeCanceled = true
 		} else {
@@ -248,7 +267,7 @@ func run() error {
 	// If Initialize sequence was canceled, don't run any other sequence.
 	if !initializeCanceled {
 		// Perform an installation if required.
-		if err = c.Run(ctx, runtime.SequenceInstall, nil); err != nil {
+		if err := c.Run(ctx, runtime.SequenceInstall, nil); err != nil {
 			return err
 		}
 
@@ -258,7 +277,7 @@ func run() error {
 		)
 
 		// Boot the machine.
-		if err = c.Run(ctx, runtime.SequenceBoot, nil); err != nil && !errors.Is(err, context.Canceled) {
+		if err := c.Run(ctx, runtime.SequenceBoot, nil); err != nil && !errors.Is(err, context.Canceled) {
 			return err
 		}
 	}
