+++ internal/app/machined/pkg/runtime/controller.go
@@ -9,6 +9,7 @@ import (
 	"log"
 
 	"github.com/cosi-project/runtime/pkg/controller"
+	"go.uber.org/zap"
 )
 
 // TaskSetupFunc defines the function that a task will execute for a specific runtime
@@ -61,4 +62,5 @@ type Controller interface {
 type V1Alpha2Controller interface {
 	Run(context.Context, *Drainer) error
 	DependencyGraph() (*controller.DependencyGraph, error)
+	MakeLogger(serviceName string) (*zap.Logger, error)
 }
